Move class selector calculation to runtime
