+++ packages/recipes/src/utils.ts
@@ -1,0 +1,12 @@
+export function mapValues<Input extends Record<string, any>, OutputValue>(
+  input: Input,
+  fn: (value: Input[keyof Input], key: keyof Input) => OutputValue,
+): Record<keyof Input, OutputValue> {
+  const result: any = {};
+
+  for (const key in input) {
+    result[key] = fn(input[key], key);
+  }
+
+  return result;
+}
